CertificateController has been updated
CertificateController delete function has been modified

certificates.delete has been created for security

certificates view files (create,index,edit,show) has been modified

+++ nk-dashboard/app/Http/Controllers/CertificateController.php
@@ -110,7 +110,11 @@ class CertificateController extends Controller
     public function destroy($id)
     {
         $certificate = Certificate::find($id);
-        $certificate->delete();
+        if ($certificate == null) {
+            return view('certificates.delete');
+        } else {
+            $certificate->delete();
+        }
         return redirect('/certificates/index');
     }
 }
